+++ exp_scenario/module/control/discord_alerts.lua
@@ -167,8 +167,8 @@ end
 
 --- When a player is jailed or unjailed
 if config.player_jail then
-    local Jail = require("modules.exp_legacy.modules.control.jail")
-    events[Jail.events.on_player_jailed] = function(event)
+    local Jail = require("modules/exp_scenario/control/jail")
+    events[Jail.on_player_jailed] = function(event)
         local player_name, by_player_name = get_player_name(event)
         emit_event{
             title = "Jail",
@@ -181,7 +181,7 @@ if config.player_jail then
             },
         }
     end
-    events[Jail.events.on_player_unjailed] = function(event)
+    events[Jail.on_player_unjailed] = function(event)
         local player_name, by_player_name = get_player_name(event)
         emit_event{
             title = "Unjail",
